requirements.txt
requests>=2.25.0
beautifulsoup4>=4.9.0
PyJWT>=2.0.0
